Updated booking-status page
