--- ride-app-front-end/src/CSS/booking/BookingDetails.css
@@ -1,164 +1,0 @@
-/* Container */
-.booking-details-container {
-  max-width: 700px;
-  margin: 4rem auto;
-  background-color: #fbfbfbf5;
-  padding: 2rem;
-  border: 1px solid #2a2a3a;
-  color: white;
-  border-radius: 8px;
-}
-
-/* Form Layout */
-.booking-details-form {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-/* Headings */
-h2,
-h3 {
-  color: var(--primary-color);
-  margin-bottom: 1rem;
-}
-
-/* Form Groups */
-.form-group {
-  display: flex;
-  flex-direction: column;
-}
-
-.form-group label {
-  margin-bottom: 0.5rem;
-  font-size: 0.85rem;
-  color: #4b4646;
-}
-
-/* Inputs & Selects */
-input,
-select,
-textarea {
-  padding: 0.75rem;
-  border-radius: 4px;
-  border: 1px solid #ccc;
-  color: #484343;
-  font-size: 1rem;
-  box-sizing: border-box;
-  background-color: white;
-}
-
-/* Checkbox Groups */
-.checkbox-group {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin: 16px 0;
-  font-family: Arial, sans-serif;
-}
-
-.checkbox-group input[type="checkbox"] {
-  width: 18px;
-  height: 18px;
-  cursor: pointer;
-  accent-color: #06375c;
-}
-
-.checkbox-group label {
-  cursor: pointer;
-  font-size: 16px;
-  color: #333;
-  margin-bottom: 0;
-}
-
-/* Full-Width Textarea */
-.full-width {
-  width: 100%;
-}
-
-/* Submit Button */
-.submit-button {
-  margin-top: 2rem;
-  background-color: var(--primary-color);
-  color: #fff;
-  padding: 0.75rem 1.5rem;
-  font-weight: bold;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.submit-button:hover {
-  background-color: var(--secondary-color);
-}
-
-.form-row {
-  display: flex;
-  gap: 1rem;
-  flex-wrap: wrap;
-}
-
-.form-row .form-group {
-  flex: 1;
-  min-width: 200px; /* prevents squishing on smaller screens */
-}
-
-.guest-card {
-  background-color: #dae7f2; /* light blue, adjust as needed */
-  padding: 1rem;
-  border-radius: 6px;
-  border: 1px solid #c9c9c9;
-  margin-bottom: 1.5rem;
-}
-
-.terms-section {
-  margin-top: 1rem;
-}
-
-.terms-section .checkbox-group {
-  margin-bottom: 0.25rem; /* Reduce vertical spacing */
-}
-
-.terms-section .checkbox-group:last-child {
-  margin-bottom: 0; /* Remove extra space after last checkbox */
-}
-
-
-.terms-link {
-  color: #0078d4;
-  text-decoration: underline;
-  cursor: pointer;
-}
-
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.6);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 999;
-}
-
-.modal-content {
-  background: #fff;
-  color: #333;
-  padding: 2rem;
-  border-radius: 8px;
-  max-width: 500px;
-  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
-}
-
-.modal-close {
-  margin-top: 1rem;
-  padding: 0.5rem 1rem;
-  background-color: var(--primary-color);
-  color: #fff;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
