--- ride-app-front-end/src/CSS/booking/BookingReview.css
@@ -1,74 +1,0 @@
-.booking-review-container {
-  max-width: 700px;
-  margin: 4rem auto;
-  padding: 2rem;
-  background-color: #f8f9fa;
-  border: 1px solid #d1d1d1;
-  border-radius: 8px;
-  font-family: "Segoe UI", sans-serif;
-  color: #333;
-}
-
-.booking-review-container h2 {
-  text-align: center;
-  color: var(--primary-color);
-  margin-bottom: 2rem;
-}
-
-.review-section {
-  background-color: #ffffff;
-  border: 1px solid #dee2e6;
-  border-radius: 6px;
-  padding: 1rem;
-  margin-bottom: 1.25rem;
-}
-
-.review-section h3 {
-  margin-bottom: 0.75rem;
-  color: var(--primary-color);
-  font-size: 1.1rem;
-}
-
-.review-section p {
-  margin: 0.4rem 0;
-  font-size: 0.95rem;
-}
-
-.review-section strong {
-  color: #222;
-}
-
-/* Button Row */
-.button-row {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 2rem;
-}
-
-.back-btn,
-.confirm-btn {
-  padding: 0.6rem 1.2rem;
-  font-weight: bold;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.back-btn {
-  background-color: #e0e0e0;
-  color: #333;
-}
-
-.back-btn:hover {
-  background-color: #d0d0d0;
-}
-
-.confirm-btn {
-  background-color: var(--primary-color);
-  color: #fff;
-}
-
-.confirm-btn:hover {
-  background-color: var(--secondary-color);
-}
