--- ride-app-front-end/src/CSS/booking/CarSelection.css
@@ -1,93 +1,0 @@
-.car-selection-container {
-  padding: 3rem;
-  max-width: 1000px;
-  margin: 0 auto;
-  color: #fff;
-  margin-top: 50px;
-}
-
-h2 {
-  font-size: 2rem;
-  margin-bottom: 2rem;
-  color: var(--primary-color);
-  text-align: center;
-}
-
-.car-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
-  gap: 1.5rem;
-  margin-bottom: 2rem;
-}
-
-.car-card {
-  background-color: #f7f7f9;
-  border: 2px solid transparent;
-  border-radius: 8px;
-  padding: 1rem;
-  cursor: pointer;
-  transition: transform 0.3s, border-color 0.3s;
-}
-
-.car-card:hover {
-  border-color: var(--primary-color);
-  transform: scale(1.02);
-}
-
-.car-card.selected {
-  border-color: var(--primary-color);
-  background-color: #262635;
-}
-
-.car-card.selected h3,
-.car-card.selected p,
-.car-card.selected .car-specs span {
-  color: white;
-}
-
-.car-image {
-  width: 100%;
-  height: 160px;
-  object-fit: cover;
-  border-radius: 4px;
-  margin-bottom: 1rem;
-}
-
-.car-specs {
-  display: flex;
-  justify-content: space-between;
-  font-size: 0.9rem;
-  margin-top: 0.5rem;
-  color: #2b2929;
-}
-
-.button-row {
-  display: flex;
-  justify-content: space-between;
-}
-
-.back-btn,
-.next-btn {
-  padding: 0.75rem 2rem;
-  border: none;
-  font-weight: bold;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 1rem;
-}
-
-.back-btn {
-  background-color: #2a2a3a;
-  color: #ddd;
-}
-
-.next-btn {
-  background-color: var(--primary-color);
-  color: white;
-}
-
-.next-btn:disabled {
-  background-color: #555;
-  color: #aaa;
-  cursor: not-allowed;
-}
