--- ride-app-front-end/src/CSS/booking/Confirmation.css
@@ -1,114 +1,0 @@
-.confirmation-container {
-  max-width: 550px;
-  margin: 4rem auto;
-  padding: 2rem;
-  background-color: #f9fbfc;
-  border: 1px solid #d5dde5;
-  border-radius: 8px;
-  font-family: "Segoe UI", sans-serif;
-  color: #333;
-  box-shadow: 0 6px 12px rgba(0, 0, 0, 0.06);
-}
-
-.confirmation-container h2 {
-  text-align: center;
-  color: var(--primary-color);
-  margin-bottom: 0.75rem;
-}
-
-.confirmation-container p {
-  text-align: center;
-  font-size: 1rem;
-  margin-bottom: 1.25rem;
-}
-
-/* Reference Box */
-.reference-box {
-  background-color: #ffffff;
-  border: 1px solid #ccd9e3;
-  border-radius: 6px;
-  padding: 1rem 1.5rem;
-  text-align: center;
-  margin-bottom: 2rem;
-}
-
-.reference-box h4 {
-  font-size: 1rem;
-  color: #06375c;
-  margin-bottom: 0.5rem;
-}
-
-.ref-number {
-  font-size: 1.4rem;
-  font-weight: bold;
-  color: #0078d4;
-  margin-bottom: 0.5rem;
-}
-
-.note {
-  font-size: 0.9rem;
-  color: #666;
-  margin-bottom: 1rem;
-}
-
-/* Steps List */
-.next-steps {
-  text-align: left;
-}
-
-.next-steps strong {
-  display: block;
-  margin-bottom: 0.5rem;
-  font-size: 1.05rem;
-  color: #333;
-}
-
-.next-steps ol {
-  padding-left: 1.25rem;
-  margin: 0;
-  font-size: 0.95rem;
-  color: #444;
-}
-
-.next-steps li {
-  margin-bottom: 0.4rem;
-}
-
-/* Copy Button */
-button {
-  display: block;
-  margin: 0 auto 2rem;
-  background-color: #06375c;
-  color: #fff;
-  padding: 0.6rem 1.2rem;
-  border: none;
-  border-radius: 4px;
-  font-weight: bold;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-button:hover {
-  background-color: #045092;
-}
-
-/* Return Link Button */
-.button-row {
-   display: flex;
-  justify-content: center; /* ⬅️ This centers the button horizontally */
-  margin-top: 2rem;
-}
-
-.return-btn {
-  background-color: var(--primary-color);
-  color: #fff;
-  padding: 0.6rem 1.4rem;
-  border-radius: 4px;
-  font-weight: bold;
-  text-decoration: none;
-  transition: background-color 0.3s ease;
-}
-
-.return-btn:hover {
-  background-color: var(--secondary-color);
-}
