--- ride-app-front-end/src/CSS/booking/Footer.css
@@ -1,70 +1,0 @@
-.footer {
-  background-color: #0a0a14;
-  color: #aaa;
-  padding: 40px 20px 20px;
-  border-top: 1px solid #2a2a3a;
-  font-size: 14px;
-}
-
-.footer-container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  max-width: 1200px;
-  margin: auto;
-  gap: 2rem;
-}
-
-.footer-col {
-  flex: 0 0 180px;
-}
-
-.footer-col h4 {
-  color: white;
-  margin-bottom: 12px;
-  text-transform: uppercase;
-  font-size: 13px;
-  letter-spacing: 1px;
-}
-
-.footer-col ul {
-  list-style: none;
-  padding: 0;
-}
-
-.footer-col ul li {
-  margin-bottom: 8px;
-}
-
-.footer-col a {
-  color: #ccc;
-  text-decoration: none;
-  transition: color 0.3s;
-}
-
-.footer-col a:hover {
-  color: #d4af37;
-}
-
-.brand {
-  font-family: "Georgia", serif;
-  color: #d4af37;
-  font-size: 20px;
-  margin-bottom: 10px;
-}
-
-.footer-bottom {
-  text-align: center;
-  margin-top: 40px;
-  padding-top: 20px;
-  border-top: 1px solid #2a2a3a;
-  font-size: 12px;
-  color: #666;
-}
-
-@media (max-width: 768px) {
-  .footer-container {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-}
