--- ride-app-front-end/src/CSS/booking/Header.css
@@ -1,31 +1,0 @@
-/* Header.css */
-
-.header {
-  background-color: transparent;
-  padding: 1rem 0;
-  position: absolute;
-  top: 0;
-  width: 100%;
-  z-index: 1000;
-}
-
-.header .container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 1.5rem;
-}
-
-/* Logo styling */
-.logo a {
-  font-size: 1.75rem;
-  font-weight: 700;
-  text-decoration: none;
-  font-family: "Georgia", serif;
-  color: #d4af37;
-  transition: color var(--transition-speed);
-}
-
-.logo a:hover {
-  color: var(--secondary-color);
-}
