+++ src/assets/css/Header.css
@@ -1,30 +1,31 @@
 /* Header.css */
+
 .header {
-  background-color: #0a0a14;
-  border-bottom: 1px solid #2a2a3a;
-  padding: 20px;
-  color: white;
+  background-color: transparent;
+  padding: 1rem 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+  z-index: 1000;
 }
-.container {
-  max-width: 1200px;
-  margin: auto;
+
+.header .container {
   display: flex;
   justify-content: space-between;
   align-items: center;
+  padding: 0 1.5rem;
 }
+
+/* Logo styling */
 .logo a {
+  font-size: 1.75rem;
+  font-weight: 700;
+  text-decoration: none;
   font-family: "Georgia", serif;
-  font-size: 24px;
   color: #d4af37;
-  text-decoration: none;
+  transition: color var(--transition-speed);
 }
-.nav a {
-  margin-left: 24px;
-  text-decoration: none;
-  color: white;
-  font-size: 14px;
-  text-transform: uppercase;
-}
-.nav a:hover {
-  color: #d4af37;
+
+.logo a:hover {
+  color: var(--secondary-color);
 }
