--- src/assets/css/booking/LandingPage.css
@@ -1,134 +1,0 @@
-/* LandingPage.css */
-
-.landing-page {
-    position: relative;
-    background-image: url("../../images/booking/singapore-flyer-219908_1280.jpg");
-    background-size: cover;
-    background-position: center;
-    height: 100vh;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    color: white;
-}
-
-.overlay {
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(
-        to right,
-        rgba(15, 15, 15, 0.9),
-        rgba(15, 15, 15, 0.4)
-    );
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 2rem;
-    box-sizing: border-box;
-}
-
-.landing-content {
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    gap: 2rem;
-    background-color: rgba(29, 28, 28, 0.75);
-    padding: 2rem;
-    border-radius: 8px;
-    color: #111;
-}
-
-.text-section {
-    flex: 1;
-    min-width: 300px;
-    max-width: 600px;
-}
-
-.text-section h1 {
-    font-size: 3rem;
-    margin-bottom: 1rem;
-    font-weight: bold;
-    color: var(--primary-dark);
-}
-
-.text-section p {
-    font-size: 1.25rem;
-    margin-bottom: 2rem;
-    color: #333;
-}
-
-.learn-more-btn {
-    background-color: var(--primary-color);
-    color: white;
-    border: none;
-    padding: 0.75rem 2rem;
-    font-size: 1rem;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color var(--transition-speed);
-}
-
-.learn-more-btn:hover {
-    background-color: var(--secondary-color);
-}
-
-.form-section {
-    flex: 1;
-    min-width: 320px;
-    max-width: 500px;
-}
-
-.landing-hero {
-    width: 100%;
-    background: url("../../images/booking/singapore-flyer-219908_1280.jpg")
-        no-repeat center center;
-    background-size: cover;
-    padding-top: 100px;
-    padding-bottom: 60px;
-}
-
-.landing-content {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    max-width: 1200px;
-    margin: 0 auto;
-    gap: 2rem;
-    padding: 2rem;
-    border-radius: 8px;
-}
-
-.hero-text {
-    flex: 1;
-    padding: 1rem;
-}
-
-.hero-text h1 {
-    font-size: 3.75rem;
-    line-height: 1;
-    color: var(--primary-dark);
-    margin-bottom: 1rem;
-    color: rgb(255 255 255);
-}
-
-.hero-text p {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    color: var(--text-light);
-    margin-bottom: 2rem;
-    color: rgb(255 255 255);
-}
-
-.learn-more-btn {
-    padding: 0.75rem 1.5rem;
-    background-color: #1e3a8a;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-.hero-form {
-    flex: 1;
-}
